Refactor and fix backup and restore
