+++ .changes/backup.md
@@ -1,0 +1,5 @@
+---
+"nodejs-binding": patch
+---
+
+Fix restore backup.
