+++ src/account_manager/operations/mod.rs
@@ -3,12 +3,12 @@
 
 pub(crate) mod account_recovery;
 pub(crate) mod background_syncing;
-#[cfg(feature = "stronghold")]
-pub(crate) mod backup;
 pub(crate) mod get_account;
 #[cfg(feature = "ledger_nano")]
 pub(crate) mod ledger_nano;
 #[cfg(feature = "stronghold")]
 pub(crate) mod stronghold;
+#[cfg(feature = "stronghold")]
+pub(crate) mod stronghold_backup;
 #[cfg(debug_assertions)]
 pub(crate) mod verify_integrity;
